Test win percentage
Add a test that ensures that in high-level play, it is sufficiently rare
for a player with 2/3rds of planets to end up losing. In this case, we
verify that we are pretty sure such scenarios are less than 1% of cases,
and can reasonably call owning 2/3rds of the planets a win.

Also add executable metadata.